feat(upgrade): Restructure update scripts (ongoing).
