+++ upgrade/old/1230-delete-previews/get-preview-paths.py
