+++ upgrade/old/1230-delete-previews/sparql/delete-preview-values.rq
