+++ upgrade/old/1230-delete-previews/sparql/get-preview-data.rq
