+++ upgrade/old/1263-knora-admin/update-knora-admin.py
