+++ upgrade/plugins/1307/knora-ontologies/knora-base.ttl
@@ -31,7 +31,9 @@
 
     rdfs:label "The Knora base ontology"@en ;
 
-    :attachedToProject knora-admin:SystemProject .
+    :attachedToProject knora-admin:SystemProject ;
+
+    :ontologyVersion "PR 1307" .
 
 
 
@@ -41,17 +43,34 @@
 #
 #################################################################
 
+
+
+### http://www.knora.org/ontology/knora-base#ontologyVersion
+
+:ontologyVersion rdf:type owl:DatatypeProperty ;
+
+    rdfs:subPropertyOf :objectCannotBeMarkedAsDeleted ;
+
+    rdfs:comment "Indicates the version of the Knora built-in ontologies in a repository."@en ;
+
+    :subjectClassConstraint owl:Ontology ;
+
+    :objectDatatypeConstraint xsd:string .
+
+
+
 ### http://www.knora.org/ontology/knora-base#isMainResource
 
 # This property is generated by a SPARQL CONSTRUCT query, it is not used in the triplestore and not used in the API
 
 :isMainResource rdf:type owl:DatatypeProperty ;
 
-    rdfs:comment "Indicates if the given resource is the main resource of a request or a resource referred to by a link property."@en;
+    rdfs:comment "Indicates if the given resource is the main resource of a request or a resource referred to by a link property."@en ;
 
     :subjectClassConstraint :Resource ;
 
     :objectDatatypeConstraint xsd:boolean .
+
 
 
 ###  http://www.knora.org/ontology/knora-base#objectCannotBeMarkedAsDeleted
