+++ upgrade/plugins/1322/knora-ontologies/knora-base.ttl
@@ -31,7 +31,9 @@
 
     rdfs:label "The Knora base ontology"@en ;
 
-    :attachedToProject knora-admin:SystemProject .
+    :attachedToProject knora-admin:SystemProject ;
+
+    :ontologyVersion "PR 1322" .
 
 
 
@@ -40,6 +42,22 @@
 #    Properties
 #
 #################################################################
+
+
+
+### http://www.knora.org/ontology/knora-base#ontologyVersion
+
+:ontologyVersion rdf:type owl:DatatypeProperty ;
+
+    rdfs:subPropertyOf :objectCannotBeMarkedAsDeleted ;
+
+    rdfs:comment "Indicates the version of the Knora built-in ontologies in a repository."@en ;
+
+    :subjectClassConstraint owl:Ontology ;
+
+    :objectDatatypeConstraint xsd:string .
+
+
 
 ### http://www.knora.org/ontology/knora-base#isMainResource
