+++ upgrade/plugins/1322/knora-ontologies/salsah-gui.ttl
@@ -1,0 +1,237 @@
+# Copyright © 2015-2019 the contributors (see Contributors.md).
+#
+# This file is part of Knora.
+#
+# Knora is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Knora is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public
+# License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
+
+@prefix : <http://www.knora.org/ontology/salsah-gui#> .
+@prefix owl: <http://www.w3.org/2002/07/owl#> .
+@prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
+@prefix xml: <http://www.w3.org/XML/1998/namespace> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+@prefix knora-base: <http://www.knora.org/ontology/knora-base#> .
+@prefix knora-admin: <http://www.knora.org/ontology/knora-admin#> .
+@base <http://www.knora.org/ontology/salsah-gui> .
+
+<http://www.knora.org/ontology/salsah-gui> rdf:type owl:Ontology ;
+
+    rdfs:label "The salsah-gui ontology"@en ;
+
+    knora-base:attachedToProject knora-admin:SystemProject .
+
+
+#################################################################
+#
+#    Properties
+#
+#################################################################
+
+
+###  http://www.knora.org/ontology/salsah-gui#guiOrder
+
+:guiOrder rdf:type owl:DatatypeProperty ;
+
+          knora-base:subjectClassConstraint owl:Restriction ;
+
+          knora-base:objectDatatypeConstraint xsd:nonNegativeInteger .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#guiElement
+
+:guiElement rdf:type owl:ObjectProperty .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#guiAttribute
+
+:guiAttribute rdf:type owl:DatatypeProperty ;
+
+          knora-base:objectDatatypeConstraint xsd:string .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#guiAttributeDefinition
+
+:guiAttributeDefinition rdf:type owl:DatatypeProperty ;
+
+                        knora-base:subjectClassConstraint :Guielement ;
+
+                        knora-base:objectDatatypeConstraint xsd:string .
+
+
+
+
+
+#################################################################
+#
+#    Classes
+#
+#################################################################
+
+
+###  http://www.knora.org/ontology/salsah-gui#Guielement
+
+:Guielement rdf:type owl:Class .
+
+
+
+
+
+#################################################################
+#
+#    Individuals
+#
+#################################################################
+
+
+###  http://www.knora.org/ontology/salsah-gui#Colorpicker
+
+:Colorpicker rdf:type :Guielement ,
+                      owl:NamedIndividual ;
+
+             :guiAttributeDefinition "ncolors:integer" .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Date
+
+:Date rdf:type :Guielement ,
+               owl:NamedIndividual .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Geometry
+
+:Geometry rdf:type :Guielement ,
+                   owl:NamedIndividual .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Geonames
+
+:Geonames rdf:type :Guielement ,
+                   owl:NamedIndividual .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Iconclass
+
+# :Iconclass rdf:type :Guielement ,
+#                     owl:NamedIndividual .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Interval
+
+:Interval rdf:type :Guielement ,
+                   owl:NamedIndividual .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#List
+
+:List rdf:type :Guielement ,
+               owl:NamedIndividual  ;
+
+          :guiAttributeDefinition "hlist(required):iri" .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Pulldown
+
+:Pulldown rdf:type :Guielement ,
+                   owl:NamedIndividual ;
+
+          :guiAttributeDefinition "hlist(required):iri" .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Radio
+
+:Radio rdf:type :Guielement ,
+                owl:NamedIndividual ;
+
+       :guiAttributeDefinition "hlist(required):iri" .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Richtext
+
+:Richtext rdf:type :Guielement ,
+                   owl:NamedIndividual .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Searchbox
+
+:Searchbox rdf:type :Guielement ,
+                    owl:NamedIndividual ;
+
+           :guiAttributeDefinition "numprops:integer" .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#SimpleText
+
+:SimpleText rdf:type :Guielement ,
+                     owl:NamedIndividual ;
+
+            :guiAttributeDefinition "size:integer" ,
+                                    "maxlength:integer" .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Slider
+
+:Slider rdf:type :Guielement ,
+                 owl:NamedIndividual ;
+
+        :guiAttributeDefinition "max(required):decimal" ,
+                                "min(required):decimal" .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Spinbox
+
+:Spinbox rdf:type :Guielement ,
+                  owl:NamedIndividual ;
+
+         :guiAttributeDefinition "max:decimal" ,
+                                 "min:decimal" .
+
+
+
+###  http://www.knora.org/ontology/salsah-gui#Textarea
+
+:Textarea rdf:type :Guielement ,
+                   owl:NamedIndividual ;
+
+          :guiAttributeDefinition "cols:integer" ,
+                                  "rows:integer" ,
+                                  "width:percent" ,
+                                  "wrap:string(soft|hard)" .
+
+
+###  http://www.knora.org/ontology/salsah-gui#Checkbox
+
+:Checkbox rdf:type :Guielement ,
+                   owl:NamedIndividual .
+
+
+###  http://www.knora.org/ontology/salsah-gui#Fileupload
+
+:Fileupload rdf:type :Guielement ,
+                     owl:NamedIndividual .
