+++ webapi/src/main/scala/org/knora/webapi/messages/v2/responder/ontologymessages/KnoraBaseToApiV2SimpleTransformationRules.scala
@@ -370,6 +370,7 @@ object KnoraBaseToApiV2SimpleTransformationRules extends OntologyTransformationR
       * Properties to remove from `knora-base` before converting it to the [[ApiV2Simple]] schema.
       */
     override val internalPropertiesToRemove: Set[SmartIri] = Set(
+        OntologyConstants.KnoraBase.OntologyVersion,
         OntologyConstants.KnoraBase.CreationDate,
         OntologyConstants.KnoraBase.LastModificationDate,
         OntologyConstants.KnoraBase.IsEditable,
